IOSystems

Basics

An input-output-system is characterized by a set of equations. These equations can be either first order ODEs or explicit algebraic equations.

\[\begin{aligned} \dot \mathbf x(t) &= f(\mathbf x(t), \mathbf y(t), \mathbf i(t), p)\\ \mathbf y(t) &= g(\mathbf x(t), \mathbf y(t), \mathbf i(t), p) \end{aligned}\]

such system contains of

states ($x$ and $y$) and

parameters ($i$, $p$).

States are determined by the given equations (i.e. the equations describe how the states change). Parameters are externally given. For IO systems we define subgroups

states

internal states (istates) which are meant for internal use

output states (outputs) which might be used as inputs for other systems

parameters

internal parameters (iparams) which are typically constant and

inputs (inputs) which can be connected to the outputs of other systems.

Types

The base type for the IOSystems is AbstractIOSystem with the has two concrete implementations.

IOSystems.AbstractIOSystem — Type

abstract supertype for IOBlock and IOSystem.

An IOBlock consists of a set of equations, a set of inputs and outputs and a name.

IOSystems.IOBlock — Type

struct IOBlock <: AbstractIOSystem

A basic IOSystem which consists of a single ODESystem.

name::Symbol

inputs::Array{SymbolicUtils.Symbolic,1}

iparams::Array{SymbolicUtils.Symbolic,1}

istates::Array{SymbolicUtils.Symbolic,1}

outputs::Array{SymbolicUtils.Symbolic,1}

system::ODESystem

removed_states::Array{SymbolicUtils.Symbolic,1}

removed_eqs::Array{Equation,1}

Missing docstring.

Missing docstring for IOBlock(eqs::AbstractVector{<:Equation}, inputs, outputs; name = gensym(:IOBlock)). Check Documenter's build log for details.

An IOSystem consists of multiple AbstractIOSystems and the connections between them.

IOSystems.IOSystem — Type

struct IOSystem <: AbstractIOSystem

A composite IOSystem which consists of multiple AbstractIOSystem which are connected via a vector of namespaced pairs (subsys1.out1 => subsys2.in1).

An IOSystem contains maps how to promote the namespaced variables of the subsystem to the new scope subsys1₊x(t) => x(t) subsys1₊y(t) => subsys1₊y(t) subsys2₊y(t) => subsys2₊y(t)

name::Symbol

inputs::Array{SymbolicUtils.Symbolic,1}

iparams::Array{SymbolicUtils.Symbolic,1}

istates::Array{SymbolicUtils.Symbolic,1}

outputs::Array{SymbolicUtils.Symbolic,1}

connections::Dict{SymbolicUtils.Symbolic,SymbolicUtils.Symbolic}

inputs_map::Dict{SymbolicUtils.Symbolic,SymbolicUtils.Symbolic}

iparams_map::Dict{SymbolicUtils.Symbolic,SymbolicUtils.Symbolic}

istates_map::Dict{SymbolicUtils.Symbolic,SymbolicUtils.Symbolic}

outputs_map::Dict{SymbolicUtils.Symbolic,SymbolicUtils.Symbolic}

systems::Array{AbstractIOSystem,1}

IOSystems.IOSystem — Method

IOSystem(cons, io_systems; inputs_map, iparams_map, istates_map, outputs_map, name, autopromote)

Construct a new IOSystem from various subsystems. Parameters

cons: the connections in the form sub2.input => sub2.output

io_systems: Vector of subsystems

inputs_map, iparams_map, istates_map: Provide custom namespace promotion / renaming i.e. sub1.input => voltage. Variables without entry in the map will be promoted automatically.

outputs_map: Provide custom namespace promotion for outputs. If this parameter is not given, all of the outputs (connected and unconnected) will be marked as outputs of the whole system. If the outputs map is given only the given outputs will be marked as outputs. All other outputs of subsystems will become internal states of the connected system (and might be optimized away in connect_system).

name: namespace

autopromote=true: enable/disable automatic promotion of variable names to system namespace

Transformations

Each IOSystem can be transformed into an IOBlock. At this step the actual manipulation of the equations happen.

IOSystems.connect_system — Function

connect_system(ios; verbose, simplify_eqs)

Recursively transform IOSystems to IOBlocks.

substitute inputs with connected outputs

try to eliminate equations for internal states which are not used to calculate the specified outputs of the system.

try to eliminate explicit algebraic equations (i.e. outputs of internal blocks) by substituting each occurrence with their rhs. Explicit algebraic states which are marked as system outputs won't be removed.

Parameters:

ios: system to connect

verbose=false: toggle verbosity (show equations at different steps)

simplify_eqs=true: toggle simplification of all equations at the end

Function building

IOSystems.generate_io_function — Function

generate_io_function(ios; first_states, first_inputs, first_params, first_removed, expression, verbose, type)

Generate callable functions for an AbstractIOSystem. An IOSystem will be transformed to an IOBlock first. At this level there is no more distinction between internal states and outputs: states=(istates ∪ outputs).

Parameters:

ios: the system to build the function

optional:

type=:auto: :ode or :static, determines the output of the function

first_states: define states=(istates ∪ outputs) which should appear first

first_inputs: define inputs which should appear first

first_params: define parameters which should appear first

expression=Val{false}: toggle expression and callable function output

Returns an named tuple with the fields

for type=:ode:

f_ip in-place function f(dstates, states, inputs, params, iv)

f_oop out-of-place function f(states, inputs, params, iv) => dstates

massm mass matrix of the system

for type=:static:

f_ip in-place function f(states, inputs, params, iv)

f_oop out-of-place function f(inputs, params, iv) => states

always:

states symbols of states (in order)

inputs symbols of inputs (in order)

params symbols of parameters (in order)