IOSystems

Types

The base type for the IOSystems is AbstractIOSystem

IOSystems.AbstractIOSystem — Type

AbstractIOSystem

Basic type for IOSystems. Such systems contains a ODEProblem of the form

\[\begin{aligned} \dot \mathbf x(t) &= f( \mathbf i(t) )\\ \mathbf o(t) &= g(\mathbf x(t), \mathbf i(t)) \end{aligned}\]

where the states can be separated into states = inputs ∪ istates ∪ outputs (istates is short for internalstates).

Each AIOS has a name field generating a namespace. The symbols of the

whith the possible concrete implementations

IOSystems.IOBlock — Type

struct IOBlock <: AbstractIOSystem

A basic IOSystem which consists of a single ODESystems.

name::Symbol

inputs::Array{SymbolicUtils.Symbolic,1}

iparams::Array{SymbolicUtils.Symbolic,1}

istates::Array{SymbolicUtils.Symbolic,1}

outputs::Array{SymbolicUtils.Symbolic,1}

system::ModelingToolkit.ODESystem

IOSystems.IOSystem — Type

struct IOSystem <: AbstractIOSystem

A composit IOSystem which consists of multipe AbstractIOSystem(@ref) wich are connected via a vector of namespaced pairs (subsys1.out1 => subsys2.in1). The inputs and outputs of the composit system are declared in the namespace of the new composit system but point to namespaced variables of the subsystems.

name::Symbol

inputs::Array{SymbolicUtils.Symbolic,1}

iparams::Array{SymbolicUtils.Symbolic,1}

istates::Array{SymbolicUtils.Symbolic,1}

outputs::Array{SymbolicUtils.Symbolic,1}

connections::Dict{SymbolicUtils.Symbolic,SymbolicUtils.Symbolic}

inputs_map::Dict{SymbolicUtils.Symbolic,SymbolicUtils.Symbolic}

iparams_map::Dict{SymbolicUtils.Symbolic,SymbolicUtils.Symbolic}

istates_map::Dict{SymbolicUtils.Symbolic,SymbolicUtils.Symbolic}

outputs_map::Dict{SymbolicUtils.Symbolic,SymbolicUtils.Symbolic}

systems::Array{AbstractIOSystem,1}